delete Offer

+++ client/src/components/layout/CardArchived.tsx
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { Offers } from "../Offer";
+import { Offers } from "../../Pages/Home";
 
 interface CardArchivedProps {
   offer: Offers;
@@ -8,19 +8,25 @@ interface CardArchivedProps {
 
 const CardArchived = ({ offer, onDelete }: CardArchivedProps) => {
   const deleteOffer = async () => {
-    try {
-      await axios.delete(
-        `${import.meta.env.VITE_OFFER_URL}/archive/${offer._id}`
-      );
-      onDelete(offer._id);
-    } catch (error) {
-      console.error("Cette offre n'a pas pu être supprimer ", error);
+    if (window.confirm("Vous voulez vraiment supprimer cette offre ?")) {
+      try {
+        await axios.delete(
+          `${import.meta.env.VITE_OFFER_URL}/archive/${offer._id}`
+        );
+        onDelete(offer._id);
+      } catch (error) {
+        console.error("Cette offre n'a pas pu être supprimer ", error);
+      }
     }
   };
 
   return (
     <div className="flex items-center justify-between rounded-md bg-orange-300 p-5 shadow-md max-lg:w-1/2 lg:w-3/4">
-      <h1 className="font-bold">{offer.title}</h1>
+      {offer.type === "Candidature spontanée" ? (
+        <h1 className="font-bold">{offer.type}</h1>
+      ) : (
+        <h1 className="font-bold">{offer.title}</h1>
+      )}
       <p>
         {offer.company} - {offer.status}
       </p>
